login/signup support for older devices
