+++ app/debug/output-metadata.json
@@ -1,0 +1,20 @@
+{
+  "version": 3,
+  "artifactType": {
+    "type": "APK",
+    "kind": "Directory"
+  },
+  "applicationId": "luci.sixsixsix.powerampache2.debug",
+  "variantName": "debug",
+  "elements": [
+    {
+      "type": "SINGLE",
+      "filters": [],
+      "attributes": [],
+      "versionCode": 27,
+      "versionName": "0.27-rc1.debug",
+      "outputFile": "app-debug.apk"
+    }
+  ],
+  "elementType": "File"
+}
