+++ app/src/main/java/luci/sixsixsix/powerampache2/presentation/main/screens/components/SheetDragHandle.kt
@@ -1,3 +1,24 @@
+/**
+ * Copyright (C) 2024  Antonio Tari
+ *
+ * This file is a part of Power Ampache 2
+ * Ampache Android client application
+ * @author Antonio Tari
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
 package luci.sixsixsix.powerampache2.presentation.main.screens.components
 
 import androidx.compose.animation.AnimatedVisibility
@@ -20,8 +41,8 @@ import androidx.compose.ui.res.dimensionResource
 import kotlinx.coroutines.launch
 import luci.sixsixsix.powerampache2.R
 import luci.sixsixsix.powerampache2.presentation.main.viewmodel.MainViewModel
-import luci.sixsixsix.powerampache2.presentation.screens_detail.song_detail.components.SongDetailTopBar
 import luci.sixsixsix.powerampache2.presentation.screens_detail.song_detail.components.MiniPlayer
+import luci.sixsixsix.powerampache2.presentation.screens_detail.song_detail.components.SongDetailTopBar
 
 @Composable
 @OptIn(ExperimentalMaterial3Api::class)
